fix: Harden input validation, memory ordering, security, and search correctness
Broad hardening commit covering input validation, ARM memory ordering,
performance, security, and search correctness across index, server, and
query components.

Input validation (index_serialization.cpp):
- Reject term_len > 10000 and posting_size > 100M in LoadFromStream
- Return kStorageCorrupted to prevent excessive allocation from malformed index files

Memory ordering (posting_list.cpp, dump_handler.cpp):
- Change doc_count_ and cached_memory_size_ stores from memory_order_relaxed
  to memory_order_release in UpdateCountsAndVersion for correct release-acquire
  pairing on ARM
- Use memory_order_release for dump_save_in_progress.store() to pair with
  existing acquire load in HandleDumpSave

Security (admin_handler.cpp):
- Remove mysql.user from CONFIG VERIFY response to prevent credential
  disclosure over TCP

Bug fixes:
- dump_handler.cpp: Remove embedded \r\n from DUMP_STARTED response which
  caused clients to truncate at the TCP frame terminator
- search_pipeline.cpp: Check ALL terms for empty n-gram results instead of
  only the first term before intersection, preventing missed early exits
- string_utils.cpp: Substitute default ngram_size=2 when kanji_ngram_size>0
  but ngram_size=0, fixing empty n-gram output in hybrid mode
- http_server.cpp: Use compare_exchange_strong in HttpServer::Stop() to
  prevent concurrent double-stop race

Performance (highlighter.cpp):
- FindMatchPositions: incremental codepoint counting from previous match
  position instead of re-scanning from byte 0 per match (O(n^2) to O(n))

Refactoring:
- string_utils.h/cpp: Widen CountCodePoints return type from uint32_t to
  size_t; add narrowing cast at call sites in highlighter.cpp
- http_server.cpp: Extract JsonFilterValueToString helper to deduplicate
  filter value coercion logic in ParseFiltersFromJson

New tests:
- tests/index/index_serialization_test.cpp: Boundary tests for term_len and
  posting_size validation in LoadFromStream
- tests/storage/document_store_retrieval_test.cpp: Correctness tests for
  GetFilterValuesBatchMultiColumn vs single-column
- tests/mysql/binlog_parsing_test.cpp: RENAME TABLE multi-pair regression tests
- tests/query/highlighter_test.cpp: Incremental codepoint counting regression
  tests for many matches and multi-byte characters
- tests/storage/document_store_docid_overflow_test.cpp: DocID rollback and
  batch duplicate handling tests
- tests/utils/string_utils_test.cpp: Regression test for ngram_size=0 with
  kanji_ngram_size>0
- tests/server/dump_handler_test.cpp: Updated assertion for removed \r\n
- tests/index/CMakeLists.txt, tests/storage/CMakeLists.txt: Register new
  test executables

+++ src/utils/string_utils.h
@@ -179,7 +179,7 @@ std::vector<std::string> GenerateQueryNgrams(std::string_view normalized, int ng
  * @param text UTF-8 encoded text
  * @return Number of Unicode code points
  */
-uint32_t CountCodePoints(std::string_view text);
+size_t CountCodePoints(std::string_view text);
 
 /**
  * @brief Check if a Unicode whitespace character starts at the given position
